requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
python-telegram-bot==20.7
APScheduler==3.10.4
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
rapidfuzz==3.6.1
python-dotenv==1.0.0
pytz==2023.3
